clarify the license for all subprojects
The LICENSE file in the root directory of the repo was a copy of the
GNU AGPL. However, the Deep User Inspector and WikiPeriod, the only
subprojects with contributions by non-maintainers, were explicitly
using the GNU GPLv3.
GeoContribs, instead, still didn't show a license header.

Thus, the LICENSE file in the root directory has been changed to use
the GNU GPLv3, and a license notice has been added to GeoContribs.